Allow higher order function spaces in outer iteration
The mixed Helmholtz problem solved in the outer iteration can be formulated in higher order function spaces, such as BDFM1 and DG1. The pressure correction equation is still solved in the lowest order space. This commit is a first implementation which runs, but the outer iteration does not converge properly yet.