Optimised implementation for diagonal lumped mass matrix
If the local 4x4 blocks of the lumped mass matrix are diagonal the implementation can be optimised both in the lumped mass matrix construction and it's application. It is only necessary to store the diagonal of the 4x4 matrix as a vector of length 4. The matrix application also simplifies as each component only has to be multiplied with the corresponding diagonal element instead of doing a 4x4 matrix multiplication.