Implemented hp-Multigrid algorithm
Wrote a class which can be used to implement a hp-cycle. The first coarsening step consists of projecting to the lowest order pressure space, on which we call the ordinary h-Multigrid. The restriction and prolongation between the low-/high-order pressure spaces are implemented in the hpMultigrid class which makes the file interordermapper.py obsolete - it is removed from the codebase.
I also adapted the driver so that it can run both at lowest and at higher order.