Loop over dof_dset of RT0 space instead of interior facet set.
Only the dof_dset's have halos, but not the interior facet set, so when building the maps in the BDFM1 lumped mass mass loop over dof_dset.set instead of interior facets. This also requires some modifications to the construction of the the diagonal matrix since we can not use the map from the interior facets to the adjacent DG0 dofs, and have to construct a map from the RT0 dof_dset instead. Code works sequentially, but currently not in parallel.