Multiply the 2nd equation in mixed system by -1 to make system (near) symmetric
If the velocity equation in the mixed system is multiplied by -1 (i.e. the RHS is defined with the opposite sign), the mixed system becomes symmetric for constant reference profile \Phi* = 1. This allows the usage of other Krylov subspace methods such as MINRES. In this commit I only changed the documentation to identify the changes which will be necessary in the code.