Use diagonal in Schur complement only
In the Schur complement preconditioner for the mixed system, add the possibility of only using the diagonal part, i.e. ignore the forward/backward substitution by the triangular matrices on either side. This might be sufficient since the Schur-complement is only used as a preconditioner so only needs to be spectrally equivalent, not necessarily exact.