Allow reading of parameters from input file.
Instead of specifying the parameters in the driver.py file, they are now read from a text file (see parameters_template.in for an example) if the name of this file is given as a command line parameter. Otherwise the defaults set in driver.py are used. The parameter parsing does some error checking (such as verifying that the type in the file matches that given by the default parameters).