Prepend local PyOP2 path, create output directory
As the code requires the local_par-loop branch of PyOP2 and the multigrid branch of firedrake, it needs to use locally installed versions of these mudules. Prepending the locations to PYTHONPATH in the job script does not seem to work for PyOP2, but this can be fixed by inserting the path at the beginning of sys.path in the python script.

Create the output directory if required.