Replace omega and omega2 by Constants to avoid recompilation.
To avoid recompilation of the code for different problem sizes, make omega and omega2 firedrake constants wherever possible. On ARCHER there is an issue with recompilation on larger core counts, so weak scaling runs will crash as the grid resolution increases on larger core counts. This workaround should ensure that the kernels only need to be compiled on smaller core counts, and are the subsequently loaded on the larger problems.