Adapted driver code and template input file to select correct mass lumping
In the input file the mass_matrix can be set to 'full', 'lump_SBR' (use mass lumping based on solid body rotation fields, which was always used so far) or 'lump_diag' (use the diagonal entries of the full velocity mass matrix to build the lumped mass matrix). For multigrid, one now also has to choose as to whether the 'diag' or 'SBR' lumping is used in the construction of the diagonal (before this was not necessary as there was only one choice for the lumped mass matrix which always has to be used there).