LU decomposition and solve
Implemented LU decomposition of the banded matrices via LAPACK. Since the code needs to be linked against the lapack and lapacke libraries, this currently requires a hack in pyop2/host.py, but in the future this could for example be passed as an additional kernel argument. Need to add further checking, in particular some mechanism of only assembling the LU matrices if they are needed.