Modified summation
With the old way of adding terms there was a rounding error of O(1E-6) when comparing the diagonal of the operator to the corresponding entries in the vertical-only matrix. This has been fixed now by slightly changing the way the expression is bracketed. Added an in-place scale operation to the banded matrix class.