Account for pointwise elimination of the buoyancy in the matrix \tilde{M}_u
In the absence of orography the buoyancy can be eliminated pointwise in the discretised equations. In this cases the matrix \tilde{M}_u simplifies considerably since it does not contain the inverse of the buoyancy mass matrix any more. Instead it becomes the velocity mass matrix and \omega_N^2 times the vertical velocity mass matrix, which can be assembled explicitly. I implemented this change and adjusted the test test_mutilde.py so that both the original setup and the setup with pointwise elimination are tested. In the absence of orography, both should give exactly the same result. However, because the field \hat{z} is discretised in the lowest order DG space, there is a difference which goes to zero as the resolution increases. To account for this, I relaxed the tolerance in the case of pointwise elimination