Allow enforcement of homogeneous Dirichlet BCs on the velocity dofs at the top and bottom of the domain
Added new method so that for a banded matrix homogeneous Dirichlet boundary conditions can be enforced at the top and bottom of the domain, i.e the boundary condition u.n=0 is assumed. This is achieved by setting the entries A_{ij} = \delta_{i,j} if i or j is a boundary node. The resulting matrix acts as the identity on the boundary dofs and has the same action on the interior dofs as the original matrix. However, we need to implement this explicitly (and can not just rely on setting the dofs to zero before and after the matrix application) whenever the banded matrix is used in the Schur complement construction.