Refactoring and generalisation of MixedArray class
In preparation for adding the more efficient solver in the absence of orography, I refactored the code such that the MixedArray, MixedOperator and MixedPreconditioner are defined in different files. I also generalised the MixedArray class so that it can be used to combine/split an arbitrary number of function spaces. These changes will make it easier to then define different operators/preconditioners for the UPB and UP systems.