Mixed outer solve without orography
In the absence of orography the buoyancy can be eliminated analytically, and it is only necessary to solve a mixed system for velocity and pressure. This can now either be done with PETSc (using Lawrence's tuned AMG preconditioner) or in a matrix-free way (for this I added two new classes which apply the U,P operator and preconditioner).