Added flexibility for lumped mass matrix
Instead of always building the lumped velocity mass matrix from div(u,v)*dx, allow building it from any UFL 2-form. This allows the generalisation to the case where the matrix \tilde{M}_u is lumped. As a first step towards implementing this I changed the interface of the LumpedMass constructor, which gets passed an UFL form instead of the function space.