Update vector assignment after PETSc upgrade
with the recent PETSc upgrade it is no longer possible to assign arrays like x.array[:]=v.array[min_idx:max_idx] if this is done more than once for the same vectors since one of the vectors reported to be in readonly state. As a workaround, copy the appropriate sections of arrays with the help of indexes.