Added class for locally assembling matrix kernels and tests
This class can be used for assembling a UFL form into local matrices without building a global matrix. This will later be used for the hybridised solver where we need those to do the local matrix inversions etc. Note that for spaces which are completely discontinuous, this representation is identical to assembling the matrix globally.