Only reassemble Mutilde if necessary.
All matrices used in the mixed U/P preconditioner are stored centrally in the class MixedOperator. If they are preassembled there, it is not necessary to assemble Mutilde again in the class Mutilde, and we can just use the matrix and it's handle from the MixedOperator class. This required adapting the code so that the Mixedoperator is build centrally and passed through the class hierarchy.