Use KSP instead of LinearSolver.solve() call to solve HDiv velocity system
In the solve() method the matrix A is applied to the boundary nodes, which is very expensive. In our case the BC nodes are always set to zero, so this is not necessary, and we can extract the KSP from the LinearSolver, use it to solve and then set the BC nodes to zero in the end. This is significantly more efficient and reduces the HDiv solve time by almost one order of magnitude	.