Use correct scaling factor in p-prologontation
Each of the fine grid (i.e. higher order) dofs should be set to the value on the coarse grid (i.e. at lowest order). For a tolerance of 1E-5 this has virtually no impact on the convergence rate (but I checked that the 1-level method shows significantly worse performance, so the coarse grid correction does make a difference).