Support single level method
So far, the multigrid level with nlevel=0 only coincides with the single level method at lowest order. At higher order a two level method is used if nlevel=0, since the p-coarsening is always used. To address this, make it possible to provide a proper single-level method. This is realised by using the Jacobi smoother as a preconditioner. Appropriate options for controlling this behaviour have been added to the parameter file.