Updated timers, consistent numbering of multigrid levels
I added timers for the application of the operators H, H_z, H_h, H_{z,inv} on each multigrid level and made the numbering of levels consistent (coarsest level is 0). Improved way of passing the multigrid level to hierarchically constructed objects, such as the smoother and the operators defined on each level).